# src/main/resources/swagger/bmp-rest.yaml
swagger: '2.0'
info:
  version: 1.0.0
  description: |
    This is a set of network protocol softwares which improve the efficiency of development.
    The purpose of efficiency improvement is not more rat race, but to get off work on time.
    Be with someone you love and/or those who love you.
  title: bmp4j integration API
host: localhost:8080
schemes:
  - http
basePath: /bmp4j
...
paths:
...
  /messages/statistic/all:
    get:
      responses:
        200:
          description: Successful Operation
          schema:
            $ref: '#/definitions/messages'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/error'

      parameters:
        - required: false
          description: The start index of paging.
          type: integer
          name: from
          in: query
        - required: false
          description: The start index of paging.
          type: integer
          name: limit
          in: query
      tags: ser-des metric
      operationId: get all statistics
      summary: show all message statistics currently.
      produces: application/json
      consumes: application/json
...
definitions:
...
  statistic:
    description: 'typical TX/RX statistic record'
    properties:
      identity:
        type: string
        description: 'messeage identity'
      receive:
        type: integer
        format: int64
        description: 'RX message count'
      transmit:
        type: integer
        format: int64
        description: 'TX message count'
      incomplete:
        type: integer
        format: int64
        description: 'unrecognized message count'
  message:
    description: 'a conncection`s serdes statistic'
    properties:
      remote:
        type: string
        description: 'remote socket like: 127.0.0.1:27492'
      local:
        type: string
        description: 'local socket like: 127.0.0.1:34819'
      statistics:
        type: array
        items:
          $ref: '#/definitions/statistic'
  messages:
    description: 'a set of message'
    properties:
      entities:
        type: array
        items:
          $ref: '#/definitions/message'
  error:
    properties:
      code:
        type: string
      message:
        type: string
      fields:
        type: string